refactor: split io pkg into multiple files
